split log in session metric hooks
